+++ src/components/ui/gameViewer/sets.js
@@ -1,6 +1,10 @@
 import { BoolEnums } from '../../../enums';
 import { ContentSwSides } from '../contentSw/enums';
 
+export const GameViewerSetEnums = {
+    gamePage: 'gamePage' // same as in textEditor, much match with Settings object
+};
+
 export const GameViewerSettings = {
     gamePage: {
         id: 'html5games',