+++ src/services/utils/arrayService.js
@@ -1,4 +1,3 @@
-import LoggerService from '../../services/loggerService';
 import { NumberService } from './numberService';
 
 export class ArrayService {
@@ -15,18 +14,5 @@ export class ArrayService {
 
     static getRandomItemFromList(list, fIndex = 0, lastIndex = list.length - 1) {
         return list[NumberService.randomInteger(fIndex, lastIndex)];
-    }
-
-    static getObject(id, obj) {
-        if (obj && obj !== undefined) {
-            return obj;
-        }
-
-        LoggerService.error(`Object #${id} is not defined`);
-        return {};
-    }
-
-    static objectPropertyAmount(obj) {
-        return obj ? Object.keys(obj).length : 0;
     }
 }