+++ src/services/utils/index.js
@@ -1,6 +1,7 @@
 export { StringService } from './stringService';
 export { ValidatorService } from './validatorService';
 export { ArrayService } from './arrayService';
+export { ObjectService } from './objectService';
 export { NumberService } from './numberService';
 export { MobileService } from '../platform/mobile';
 export { JSONService } from '../storage/jsonService';