+++ Changelog.md
@@ -1,6 +1,7 @@
 ## Changelog
 v 2.60
-  Added Feature list improvements
+  Added Feature list improvements,
+  Added GAmeViewer settings same as in TextEditor settings 
 
 v 2.59 
   Added GameViewerHelper, added TextEditor improvements